Refactor to abstract out duplication in the Meta classes
